Roll up the entire <table> element into ExpressionTable.
